dex/networks/eth: nuke coin types use common one
